build: Remove unnecessary linkages to shared libraries.
The core does not need to be linked with the ae-bridge because it does
not invoke any ae-bridge methods.

None of the plugins needed to be linked with the core because they do
not invoke core methods; they merely handle callbacks.

The front-end does not need to be linked with the core because it dyn-
amically loads the core library at run-time.