Disable extremely verbose logging.
The LOG(...) macro should only be used for debugging because it imposes
a significant performance penalty that is very noticeable on mobile
devices.

For example, on a Tegra3-based 2012 Nexus 7, framerate during the DK64
intro scene/rap is very noticeably slower with LOG enabled:

   with LOG enabled: 7-20 FPS
   with LOG disabled: 15-30 FPS

In many places in the scene, FPS doubles after disabling the LOG macro.